Big update to readme file with instructions
Important change that turns off convex hulling by default, because Rebecca Lowes discovered if one hull is inside the other, it doesn't count as a collision.
I guess that could still be the case, but far less likely, for the meshes themselves.

It's now an option in the UI, and the default is off.  You can use it, but your range should be realistic to start with.